Update authorization in CalendarInterface

+++ src/main/java/edu/cmu/sv/app17/rest/CalendarInterface.java
@@ -5,7 +5,9 @@ import com.mongodb.MongoSocketOpenException;
 import edu.cmu.sv.app17.exceptions.APPBadRequestException;
 import edu.cmu.sv.app17.exceptions.APPInternalServerException;
 import edu.cmu.sv.app17.exceptions.APPNotFoundException;
+import edu.cmu.sv.app17.exceptions.APPUnauthorizedException;
 import edu.cmu.sv.app17.helpers.APPResponse;
+import edu.cmu.sv.app17.helpers.Authorization;
 import edu.cmu.sv.app17.helpers.PATCH;
 import com.fasterxml.jackson.databind.ObjectWriter;
 import com.mongodb.BasicDBObject;
@@ -21,6 +23,8 @@ import org.bson.types.ObjectId;
 import org.json.JSONException;
 
 import javax.ws.rs.*;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import java.text.DateFormat;
 import java.text.ParseException;
@@ -47,7 +51,7 @@ public class CalendarInterface {
 
     @GET
     @Produces({ MediaType.APPLICATION_JSON})
-    public APPResponse getAll(@DefaultValue("_id") @QueryParam("sort") String sortArg) {
+    public APPResponse getAll(@Context HttpHeaders headers, @DefaultValue("_id") @QueryParam("sort") String sortArg) {
 
         ArrayList<Calendar> calList = new ArrayList<Calendar>();
 
@@ -58,6 +62,7 @@ public class CalendarInterface {
         });
 
         try {
+            Authorization.checkAdmin(headers);
             SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
             FindIterable<Document> results = collection.find().sort(sortParams);
             for (Document item : results) {
@@ -72,17 +77,17 @@ public class CalendarInterface {
             }
             return new APPResponse(calList);
 
-        } catch(Exception e) {
-            System.out.println("EXCEPTION!");
-            e.printStackTrace();
-            throw new APPInternalServerException(99,e.getMessage());
+        } catch(APPUnauthorizedException e){
+            throw new APPUnauthorizedException(70,"Not authorized.");
+        }  catch(Exception e) {
+            throw new APPInternalServerException(99,"Something happens!");
         }
     }
 
     @GET
     @Path("{id}")
     @Produces({ MediaType.APPLICATION_JSON})
-    public APPResponse getOne(@PathParam("id") String id) {
+    public APPResponse getOne(@Context HttpHeaders headers,@PathParam("id") String id) {
 
 
         BasicDBObject query = new BasicDBObject();
@@ -94,6 +99,7 @@ public class CalendarInterface {
             if (item == null) {
                 throw new APPNotFoundException(0, "Cal not found.");
             }
+            Authorization.checkSelfandAdmin(headers,item.getString("userId"));
             Calendar cal = new Calendar(
                     item.getString("userId"),
                     sdf.format(item.getDate("date")),
@@ -104,8 +110,10 @@ public class CalendarInterface {
 
         } catch(IllegalArgumentException e) {
             throw new APPBadRequestException(45,"Doesn't look like MongoDB ID");
+        }  catch(APPUnauthorizedException e){
+            throw new APPUnauthorizedException(70,"Not authorized.");
         }  catch(Exception e) {
-            throw new APPInternalServerException(99,"Something happened, pinch me!");
+            throw new APPInternalServerException(99,"Something happens!");
         }
 
 
@@ -115,7 +123,7 @@ public class CalendarInterface {
     @Path("{id}")
     @Consumes({ MediaType.APPLICATION_JSON})
     @Produces({ MediaType.APPLICATION_JSON})
-    public APPResponse update(@PathParam("id") String id, Object request) throws ParseException {
+    public APPResponse update(@Context HttpHeaders headers,@PathParam("id") String id, Object request) throws ParseException {
         JSONObject json = null;
         try {
             json = new JSONObject(ow.writeValueAsString(request));
@@ -125,7 +133,7 @@ public class CalendarInterface {
         }
 
         try {
-
+            Authorization.checkAdmin(headers);
             BasicDBObject query = new BasicDBObject();
             query.put("_id", new ObjectId(id));
 
@@ -150,6 +158,10 @@ public class CalendarInterface {
         } catch(JSONException e) {
             System.out.println("Failed to update a cal");
 
+        } catch(APPUnauthorizedException e){
+            throw new APPUnauthorizedException(70,"Not authorized.");
+        }  catch(Exception e) {
+            throw new APPInternalServerException(99,"Something happens!");
         }
         return new APPResponse();
     }
@@ -157,13 +169,20 @@ public class CalendarInterface {
     @DELETE
     @Path("{id}")
     @Produces({ MediaType.APPLICATION_JSON})
-    public APPResponse delete(@PathParam("id") String id) {
+    public APPResponse delete(@Context HttpHeaders headers,@PathParam("id") String id) {
         BasicDBObject query = new BasicDBObject();
         query.put("_id", new ObjectId(id));
+        try {
+            Authorization.checkAdmin(headers);
 
-        DeleteResult deleteResult = collection.deleteOne(query);
-        if (deleteResult.getDeletedCount() < 1)
-            throw new APPNotFoundException(66,"Could not delete");
+            DeleteResult deleteResult = collection.deleteOne(query);
+            if (deleteResult.getDeletedCount() < 1)
+                throw new APPNotFoundException(66, "Could not delete");
+        } catch(APPUnauthorizedException e){
+            throw new APPUnauthorizedException(70,"Not authorized.");
+        }  catch(Exception e) {
+            throw new APPInternalServerException(99,"Something happens!");
+        }
 
         return new APPResponse();
     }
